Add pull-down resistors for ethernet biasing networks
Without knowing the internal schematics of the Ethernet PHY driver
circuitry, we cannot assume that biasing to 3.3V is acceptable for all
PHYs involved.  For a number of PHYs, biasing must be done to 2.1V, for
example.  Hence, we need full resistive dividers to be able to set any
bias voltage as needed.  Actual resistor values TBD by experiments.